test(tokens): move the render tests beside the renderer
Task 1 left them in the checkpoint command's test file to keep a
zero-behaviour-change refactor free of churn. The renderer now has two
callers and its own file, so its tests follow. The checkpoint report and
weights tests stay: this is a split, not a move.

TestUnpricedReasons_AreScopeNeutral goes to a third file, token_cost_test.go,
beside TestTokenClassShares_UnpricedReasonsAreSpecific: it is a pure-constant
test about a token_cost.go const block, so it belongs neither in the live
path's file nor with the renderer. Two tests about one const block in two
files is the same misfiling in a new place.

Verified no test was lost or duplicated: the plan's RUN count over the
affected prefixes is 62 before and after.
